+++ icegraph/data/merge/__init__.py
@@ -1,2 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
+
+from .models import LMDBMerger, HDF5Merger
+
+__all__ = ["LMDBMerger", "HDF5Merger"]