+++ icegraph/data/split/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import SplitFactory
+from .models import DatasetSplitter
 
-__all__ = ["SplitFactory"]
+__all__ = ["DatasetSplitter"]