+++ icegraph/data/transform/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import GraphSamplePreprocessor
+from .models import TransformToDataset
 from .schemas import generate_vector_mapping
 
-__all__ = ["GraphSamplePreprocessor", "generate_vector_mapping"]
+__all__ = ["TransformToDataset", "generate_vector_mapping"]